# chore: Optimize Vector to binary perf and vice-versa